auth_service/requirements.txt
flask
flask-sqlalchemy
psycopg2-binary
pyjwt
bcrypt
python-dotenv
